+++ Formula/gomoufox.rb
@@ -1,13 +1,13 @@
 class Gomoufox < Formula
   desc "Go driver, CLI, and MCP server for Camoufox"
   homepage "https://github.com/ehmo/gomoufox"
-  version "0.1.10"
+  version "0.1.11"
   license "MIT"
 
   on_macos do
     if Hardware::CPU.arm?
-      url "https://github.com/ehmo/gomoufox/releases/download/v0.1.10/gomoufox_0.1.10_darwin_arm64.tar.gz"
-      sha256 "e65614ae61533e8fd323340ca433d46ebff47c12750002abbe6f440048d42377"
+      url "https://github.com/ehmo/gomoufox/releases/download/v0.1.11/gomoufox_0.1.11_darwin_arm64.tar.gz"
+      sha256 "434226e6b74dc853659e82cf2a843394e7e2af8ddb17544c643e47bd35aa9794"
     else
       odie "gomoufox Homebrew requires Apple Silicon because pinned Camoufox has no supported macOS Intel browser binary"
     end
@@ -17,8 +17,8 @@ class Gomoufox < Formula
     if Hardware::CPU.arm?
       odie "gomoufox Homebrew requires Linux amd64 because pinned Camoufox has no supported Linux ARM browser binary"
     else
-      url "https://github.com/ehmo/gomoufox/releases/download/v0.1.10/gomoufox_0.1.10_linux_amd64.tar.gz"
-      sha256 "7ac29d7769093a9fdad069b4e1d0b4530a55afe8e948db6c274e7b67aa078998"
+      url "https://github.com/ehmo/gomoufox/releases/download/v0.1.11/gomoufox_0.1.11_linux_amd64.tar.gz"
+      sha256 "5f3312fe6da04675d81167a845317428ebfde754d02192c18a687bf1a44713ae"
     end
   end